+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## 1.0.0 - Uncommitted
+## 1.0.1 - 2020-04-13
+### Fixed
+- Fixed bug where extending a logic with stored values would re-store the values
+- Fixed bug where using in .extend() would require a path to be given again 
+
+## 1.0.0 - 2019-09-12
 ### Changed
 - Works with kea 1.0.
 - Use `localStoragePlugin.configure(storageEngine)` to provide an alternative storage engine